components/module-market/src/main/java/com/zkdlu/order/OrderApi.java
package com.zkdlu.order;

import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

@RequiredArgsConstructor
@RestController
public class OrderApi {
    private final OrderService orderService;

    @ResponseStatus(HttpStatus.CREATED)
    @PostMapping("/order/{productId}")
    public OrderResponse placeOrder(@PathVariable Long productId) {
        return orderService.placeOrder(productId);
    }
}
